requirements.txt
torch>=1.13.0
torchvision>=0.14.0
scikit-learn>=1.0
pandas>=1.3.0
numpy>=1.21
matplotlib
seaborn
streamlit>=1.10
Pillow>=8.0
joblib

# Geospatial + OSM
geemap
earthengine-api
osmnx>=1.5.1
shapely>=1.7
folium
gdown
leafmap
streamlit-folium

# Web/IO
requests
tqdm


